🧹 cleaning up the code
